Move models out into own module
